// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        KustoQueryRun: {
            /** Executedatetimeutc */
            executeDateTimeUtc: string;
            executionMetrics?: components["schemas"]["KustoQueryStats"] | null;
            /** Expiresat */
            expiresAt: string;
            kustoQuery: components["schemas"]["KustoQueryEcho"];
            /** Mainerror */
            mainError?: string | null;
            /**
             * Queryrunid
             * Format: uuid
             */
            queryRunId: string;
            /** Resultdata */
            resultData?: {
                [key: string]: unknown;
            }[] | null;
            status: components["schemas"]["QueryRunStatus"];
        };
// ...
        ProblemDetails: {
            /** Detail */
            detail: string;
            /**
             * Errors
             * @default null
             */
            errors: {
                [key: string]: string[];
            } | null;
            /** Status */
            status: number;
            /** Title */
            title: string;
            /** Traceid */
            traceId: string;
            /** Type */
            type: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}